+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.2.0] - 2025-04-04
+
+### Changed
+- Add support for SSE and Websocket remote MCP servers
+- Introduced `McpServersConfig` type
+- Changed `stderr` of `McpServersConfig` to `errlog` to follow Python SDK more closely
+- Use double quotes instead of single quotes for string literals
+
+
 ## [0.1.11] - 2025-03-31
 
 ### Changed